# docs/swagger.yaml
basePath: /
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  models.Page:
    properties:
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updatedAt:
        type: string
    required:
    - content
    type: object
...
  models.ResponseMessage:
    properties:
      code:
        example: 302
        type: integer
      message:
        example: message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a 51cloudclass.com web api server.
  title: 51cloudclass Gin Swagger API
  version: 1.0.0
paths:
  /api/page/all:
    get:
      consumes:
      - application/json
      description: List all pages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Page'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: List all pages
      tags:
      - page
  /api/page/new:
    post:
      consumes:
      - application/json
      description: Create a new page and post to server
      parameters:
      - description: Page with content
        in: body
        name: page
        required: true
        schema:
          $ref: '#/definitions/models.Page'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseMessage'
      summary: Post a new Page
      tags:
      - page
...
schemes:
- http
swagger: "2.0"